Rename SubscriberWithResult to CompletionSubscriber
Fixes #5

As discussed in the MicroProfile Reactive hangout, this class is-a
Subscriber, rather than has-a subscriber. This makes the API
somewhat more fluent, since the naming makes it sound like its a
subscriber. Additionally, it's been made a final class, with the only
way to instantiate it being the `of` method, this is to ensure end users
don't get confused into implementing it themselves.

Signed-off-by: James Roper <james@jazzy.id.au>