Define onError, onTerminate and onComplete
Add the 3 stages called at 3 different moments of the stream lifecycle. onError is called when the stream conveys a failure. onComplete is called when the stream completes. onTerminate is called in both cases.
onComplete is already the equivalent of the BaseStream.onClose method.

The TCK for the 3 stages is merged in a single file as these stages behaves almost indentically. The only difference is the "when" the passed functions are called.

Signed-off-by: Clement Escoffier <clement.escoffier@gmail.com>