Added support for running the TCK in Arquillian
This provides a test class that can be run in Arquillian. I have tested
this with Thorntail (using an Arquillian AuxillaryArchiveAppender to add
in my own implementation), and it works.

I did originally try and make this work better with Arquillian, by
trying to have it deploy each test to the container. I was able to work
around the fact that the tests are created by a factory by writing my
own custom container test runner that gets them from the factory. I was
able to work around the fact that none of the tests extend the TestNG
Arquillian by extracting all its logic into a TestNG listener that was
applied to the whole suite. And finally, I was able to work around the
fact that none of the test classes has an @Deployment annotated method
on it, by writing a custom DeploymentScenarioGenerator that extracted
the deployment from a different class - the same class for each test.
This was the most annoying one because WildFly has its own custom one,
so doing this in a generic way was impossible. So, in the end it
worked, however, it was a lot of what felt like very fragile code, so my
strategy in the end was simply to write one Arquillian test that has one
test method, and then, in that test method, instantiate a TestNG runner
that runs the entire suite for me, using a CDI injected
ReactiveStreamsEngine. If you run this from the IDE, it says there's
only one test, but the logging shows 1100 or so tests, and makes it
clear which failed and what the failures are.

Also fixed two bugs in the TCK where the default engine was loaded
instead of the passed in one.