Generate marble diagrams as part of build
As discussed in the MicroProfile Reactive weekly hangout, I've deleted
all the marble diagram PNGs from the repo, and instead ensured they are
built by Maven. This is using the Maven frontend plugin to download and
install node and npm, and then running the script using npm. The
diagrams are generated at the last possible minute before they are
needed, that is, in the prepare-package phase.