Generate diagrams in marble-diagrams profile
This modifies the build so that diagrams are checked into git, and are
only generated when the marble-diagrams profile is activated, so that
the CI and releasing can be done on the MicroProfile build server
without needing the Chrome dependencies.

Additionally, a verification task has been added that tracks hashes of
all the input and output files to ensure that if someone changes the
diagrams but doesn't regenerate them, the build fails.