Renamed spi.CompletionSubscriber and removed hasInlet/Outlet
I probably should have done these changes in separate commits.

The first change renames spi.CompletionSubscriber to
SubscriberWithCompletionStage. Having both CompletionSubscriber in the
API and in the SPI makes things very awkward and confusing for
implementations, so I've renamed it to SubscriberWithCompletionStage in
the SPI, to differentiate between the two classes.

The second change removes Graph.hasInlet/hasOutlet and
Stage.hasInlet/hasOutlet. These serve no functional purpose, for stages,
the booleans they return can be inferred by the type of stage, and for
graphs, the information can be inferred by inspecting the graph. They
were used primarily for validation of the graph that the builder was
building, but since the builder is in complete control of the building,
and its implementation ensures that it never builds invalid graphs, this
validation doesn't serve any useful purpose beyond detecting fairly
trivial bugs.

Some implementations of the SPI may depend on these methods, but they
can easily obtain the information by inspecting the stage/graph.