Split operators into multiple interfaces
This is to group them for clarity. One advantage here is that now the
interface for Publisher/Processor builder is shared, so there's no fear
of getting it wrong when copying, and no need to double maintain the
javadocs.

One thing I'm not sure about, since we need to override all the methods
to ensure the correct subtype is returned by the builders, will the IDEs
group the methods as intended? Or do we have to introduce a second layer
of interfaces for each of publisher and processor to do that?

Also fixed a number of variance bugs, eg to/via all had no variance on
the subscriber/processor type arguments, and methods like
flatMapRsPublisher needed to be not just covariant on the elements in
the type, but covariant on the Publisher itself too.

One thing that I haven't made contravariant is functions that accept a
`Throwable`. If they were contravariant, you could pass a function that
accepted `Object` to them. But is that desired, or do we want to
explicitly force functions to accept a `Throwable` and nothing wider?
`CompletionStage` is not contravariant on its `Throwable` consuming
functions, so I thought consistency with that is probably a good idea.